backend/classifier/requirements.txt
fastapi
uvicorn
tensorflow
python-multipart
Pillow
numpy
ezkl
pyyaml
uuid
asyncio
